MsbmsSystAuthentication (msbms_syst_authentication v0.1.0)

API for the management of system user authentication.

Authentication is a global concern allowing users to be either bound to specific Application Instances and Owners or to be independent and valid for a number of Applications and Instances.

API - Access Accounts

create_access_account(access_account_params)

@spec create_access_account(MsbmsSystAuthentication.Types.access_account_params()) ::
  {:ok, MsbmsSystAuthentication.Data.SystAccessAccounts.t()}
  | {:error, MsbmsSystError.t()}

Create a new Access Account.

Parameters

access_account_params - a map of required and optional parameters which define the new Access Account record. Required attributes in this map are: internal_name, external_name, access_account_state_id, and allow_global_logins.

Example

iex> state = MsbmsSystEnums.get_default_enum_item("access_account_states")
iex> {:ok, %MsbmsSystAuthentication.Data.SystAccessAccounts{}} =
...>   MsbmsSystAuthentication.create_access_account(
...>     %{
...>       internal_name: "example_create_accnt",
...>       external_name: "Create Access Account Example",
...>       access_account_state_id: state.id,
...>       allow_global_logins: true
...>      }
...>    )

get_access_account_by_name(access_account_name)

@spec get_access_account_by_name(MsbmsSystAuthentication.Types.access_account_name()) ::
  MsbmsSystAuthentication.Data.SystAccessAccounts.t()
  | {:error, MsbmsSystError.t()}

Retrieves a fully populated Access Account record as found by internal name.

'Fully populated' in this context means that the related Account Account State and related functional type data is also retrieved along with the actual Access Account data.

Parameters

access_account_name - the internal name of the Access Account record to retrieve.

Example

iex> {
...>   :ok,
...>   %MsbmsSystAuthentication.Data.SystAccessAccounts{internal_name: "example_accnt"}
...> } =
...>   MsbmsSystAuthentication.get_access_account_by_name("example_accnt")

get_access_account_id_by_name(access_account_name)

@spec get_access_account_id_by_name(
  MsbmsSystAuthentication.Types.access_account_name()
) ::
  {:ok, MsbmsSystAuthentication.Types.access_account_id()}
  | {:error, MsbmsSystError.t()}

Looks up an Access Account record ID by its internal name.

Parameters

access_account_name - the internal name of the record for which to retrieve the record ID.

Examples

iex> {:ok, access_account_id}  =
...>   MsbmsSystAuthentication.get_access_account_id_by_name("example_accnt")
iex> is_binary(access_account_id)
true

purge_access_account(access_account)

@spec purge_access_account(
  MsbmsSystAuthentication.Types.access_account_id()
  | MsbmsSystAuthentication.Data.SystAccessAccounts.t()
) :: :ok | {:error, MsbmsSystError.t()}

Purges the requested Access Account if the Access Account State is of a purge eligible functional type.

Parameters

access_account - is either the record ID of the Access Account to purge or the populated MsbmsSystAccessAccount.Data.SystAccessAccounts struct representing the record to purge.

Example

iex> {:ok, target_access_account} =
...>   MsbmsSystAuthentication.get_access_account_by_name("example_purge_accnt")
iex> MsbmsSystAuthentication.purge_access_account(target_access_account)
:ok

update_access_account(access_account, access_account_params)

@spec update_access_account(
  MsbmsSystAuthentication.Types.access_account_id()
  | MsbmsSystAuthentication.Data.SystAccessAccounts.t(),
  MsbmsSystAuthentication.Types.access_account_params()
) ::
  {:ok, MsbmsSystAuthentication.Data.SystAccessAccounts.t()}
  | {:error, MsbmsSystError.t()}

Updates the maintainable fields of a given Access Account record.

Parameters

access_account - either the record ID value of the Access Account to update or is the complete MsbmsSystAccessAccount.Data.SystAccessAccounts struct representing the before-update state of the Access Account record.

access_account_params - a map containing those attributes to be changed along with their new values.

Example

iex> {:ok, target_access_account} =
...>   MsbmsSystAuthentication.get_access_account_by_name("example_accnt")
iex> {:ok, updated_access_account} =
...>   MsbmsSystAuthentication.update_access_account(
...>     target_access_account,
...>     %{external_name: "Updated Example Account Name"}
...>   )
iex> %MsbmsSystAuthentication.Data.SystAccessAccounts{
...>   external_name: "Updated Example Account Name"
...> } = updated_access_account

API - Access Account Instance Assocs

accept_instance_invite(access_account_instance_assoc)

@spec accept_instance_invite(
  MsbmsSystAuthentication.Types.access_account_instance_assoc_id()
  | MsbmsSystAuthentication.Data.SystAccessAccountInstanceAssocs.t()
) ::
  {:ok, MsbmsSystAuthentication.Data.SystAccessAccountInstanceAssocs.t()}
  | {:error, MsbmsSystError.t()}

Accepts the invitation made to an Access Account to access an Instance.

This process is only needed in the case where a syst_access_account_instance_assocs invitation may be accepted or declined by the Access Account holder. To be accepted in this process the invitation record must have been previously created (see invite_to_instance/3), must not be expired, previously declined, or previously accepted; trying to accept a record in such a state will result in an error tuple being returned.

Naturally, the Access Account holder may choose to never accept or decline the invitation and may simply allow the invitation to expire.

Parameters

access_account_instance_assoc - this value may be either the populated MsbmsSystAuthentication.Data.SystAccessAccountInstanceAssocs struct to accept or the record ID of the record to accept.

accept_instance_invite(access_account_id, instance_id)

@spec accept_instance_invite(
  MsbmsSystAuthentication.Types.access_account_id(),
  MsbmsSystInstanceMgr.Types.instance_id()
) ::
  {:ok, MsbmsSystAuthentication.Data.SystAccessAccountInstanceAssocs.t()}
  | {:error, MsbmsSystError.t()}

Accepts the invitation made to an Access Account to access an Instance, referencing the record by its composite key values.

This function performs the same process as accept_instance_invite/1, but will look up the record to accept using the Access Account, Instance, and Credential Type record ID values.

Parameters

access_account_id - the record ID of the Access Account accepting the invitation to the Instance.

instance_id - the Instance record ID to which the Access Account has been invited.

decline_instance_invite(access_account_instance_assoc)

@spec decline_instance_invite(
  MsbmsSystAuthentication.Types.access_account_instance_assoc_id()
  | MsbmsSystAuthentication.Data.SystAccessAccountInstanceAssocs.t()
) :: :ok | {:error, MsbmsSystError.t()}

Declines an unaccepted/unexpired invitation made to an Access Account to access an Instance.

This process is only needed in the case where a syst_access_account_instance_assocs invitation may be accepted or declined by the Access Account holder. To be declined in this process the invitation record must have been previously created (see invite_to_instance/3), must not be expired, previously declined, or previously accepted; trying to decline a record in such a state will result in an error tuple being returned.

Naturally, the Access Account holder may choose to never accept or decline the invitation and may simply allow the invitation to expire.

Parameters

access_account_instance_assoc - this value may be either the populated MsbmsSystAuthentication.Data.SystAccessAccountInstanceAssocs struct to decline or the record ID of the record to decline.

decline_instance_invite(access_account_id, instance_id)

@spec decline_instance_invite(
  MsbmsSystAuthentication.Types.access_account_id(),
  MsbmsSystInstanceMgr.Types.instance_id()
) ::
  {:ok, MsbmsSystAuthentication.Data.SystAccessAccountInstanceAssocs.t()}
  | {:error, MsbmsSystError.t()}

Declines an unaccepted/unexpired invitation made to an Access Account to access an Instance, referencing the record by its composite key values.

This function performs the same process as decline_instance_invite/1, but will look up the record to decline using the Access Account, Instance, and Credential Type record ID values.

Parameters

access_account_id - the record ID of the Access Account declining the invitation to the Instance.

instance_id - the Instance record ID to which the Access Account has been invited.

invite_to_instance(access_account_id, instance_id, opts \\ [])

@spec invite_to_instance(
  MsbmsSystAuthentication.Types.access_account_id(),
  MsbmsSystInstanceMgr.Types.instance_id(),
  Keyword.t()
) ::
  {:ok, MsbmsSystAuthentication.Data.SystAccessAccountInstanceAssocs.t()}
  | {:error, MsbmsSystError.t()}

Invites or re-invites an Access Account to establish access rights to a specific Instance.

The invitation process creates a syst_access_account_instance_assocs record for the requested Access Account, Instance, and Credential Type and sets the invitation_issued field to the current date/time. Re-invitation will update an existing syst_access_account_instance_assocs record resetting invite data such as the invitation expiration date or resetting the declined state if the record was previously declined by the Access Account holder. Once a syst_access_account_instance_assocs record is accepted, it may not be re-invited again using this process until the record is revoked (deleted).

syst_access_account_instance_assocs are unique to the Access Account and Instance combination and only one such record may exist for that combination at any one time.

Parameters

access_account_id - the record ID of the Access Account to invite to the Instance.

instance_id - the Instance record ID to which the Access Account is being invited.

opts - a keyword list of optional parameters used to set record values and behaviors. Available options are:

create_accepted - will create a syst_access_account_instance_assocs record which is already accepted when set true. The default value is false.

expiration_days - for records that are not created as accepted and act as true invitations, this option sets the number of days in which an Access Account holder may accept the invitation. After this time the invitation is considered expired and must be re-invited before it can be used to grant access again. The default value for this option is 30 days.

revoke_instance_access(access_account_instance_assoc)

@spec revoke_instance_access(
  MsbmsSystAuthentication.Types.access_account_instance_assoc_id()
  | MsbmsSystAuthentication.Data.SystAccessAccountInstanceAssocs.t()
) :: :ok | {:error, MsbmsSystError.t()}

Revokes the access or invitation to access an Instance from the given Access Account.

Simply put, Access Accounts are both invited and granted access to Instances via Access Account Instance Association records (see: MsbmsSystAuthentication.Data.SystAccessAccountInstanceAssocs) and this function deletes those records. This has the effect of revoking the invitation to access an Instance from the Access Account/Credential Type combination.

Records in any state of invited, accepted, or declined may be revoked/deleted using this function.

Parameters

access_account_instance_assoc - this value may be either the populated MsbmsSystAuthentication.Data.SystAccessAccountInstanceAssocs struct to revoke or the record ID of the record to revoke.

revoke_instance_access(access_account_id, instance_id)

@spec revoke_instance_access(
  MsbmsSystAuthentication.Types.access_account_id(),
  MsbmsSystInstanceMgr.Types.instance_id()
) ::
  {:ok, MsbmsSystAuthentication.Data.SystAccessAccountInstanceAssocs.t()}
  | {:error, MsbmsSystError.t()}

Revokes the access or invitation to access an Instance from the given Access Account/Credential Type combination, referencing the record by its composite key values.

This function performs the same action as revoke_instance_access/1 but identified the record being revoked (deleted) using its composite/candidate key values.

Parameters

access_account_id - the record ID of the Access Account from which the invitation to access is being revoked.

instance_id - the Instance record ID from which the Access Account had previously been invited to access.

API - Password Rules

create_disallowed_password(password)

@spec create_disallowed_password(MsbmsSystAuthentication.Types.credential()) ::
  :ok | {:error, MsbmsSystError.t()}

Adds a new password to the Disallowed Passwords list.

Disallowed passwords are passwords that are commonly known, are known to have been revealed in a successful hacking attack, or are otherwise not available for users to choose for their authentication credential. Enforcing that these passwords are not available for use depends upon the effective disallow_compromised Password Rule for the Access Account attempting authentication.

An attempt to add a password which is already on the list will succeed as though the password were not already part of the list.

Parameters

password - The plaintext password to add to the list of disallowed passwords.

Examples

Adding a password successfully will simply return :ok.

iex> MsbmsSystAuthentication.create_disallowed_password("Example Disallowed Password")
:ok

Any subsequent attempt to add the same password to the list again will appear to succeed while silently doing nothing.

iex> MsbmsSystAuthentication.create_disallowed_password("Example Disallowed Password")
:ok

delete_disallowed_password(password)

@spec delete_disallowed_password(MsbmsSystAuthentication.Types.credential()) ::
  {:ok, :deleted | :not_found} | {:error, MsbmsSystError.t()}

Removes a password from the disallowed passwords list.

On success, this function will return a success tuple indicating if the requested password was deleted from the disallowed passwords list ({:ok, :deleted}) or if the password simply wasn't found in the list ({:ok, :not_found}).

Parameters

password - the plaintext password to delete from the disallowed passwords list.

Examples

Result when deleting a record from the list.

iex> MsbmsSystAuthentication.delete_disallowed_password("No Longer Disallowed")
{:ok, :deleted}

Result when trying to delete a record not already on the list.

iex> MsbmsSystAuthentication.delete_disallowed_password("Not on List")
{:ok, :not_found}

password_disallowed(password)

@spec password_disallowed(MsbmsSystAuthentication.Types.credential()) ::
  {:ok, boolean()} | {:error, MsbmsSystError.t()}

Indicates whether the requested password is disallowed.

This function returns a tuple in the form of {:ok, <disallowed>} where the disallowed value is either true meaning that the requested password is disallowed or false if the password is available for use.

Regardless of the return of this function, disallowed passwords are only prevented for use if the effective disallow_compromised Password Rule for the Access Account attempting authentication is set.

Parameters

password - the plaintext password to test for disallowed status.

Examples

When a password has been previously disallowed and cannot be used as a user credential.

iex> MsbmsSystAuthentication.password_disallowed("Is Disallowed")
{:ok, true}

When a password has not been previously disallowed.

iex> MsbmsSystAuthentication.password_disallowed("Is Not Disallowed")
{:ok, false}

password_disallowed?(password)

@spec password_disallowed?(MsbmsSystAuthentication.Types.credential()) :: boolean()

Indicates whether the requested password is disallowed, raising on error.

This function works the same as disallowed_password/1 except this function returns a simple boolean value rather than a result tuple. If an error is encountered an exception is raised.

Parameters

password - the plaintext password to test for disallowed status.

Examples

When a password has been previously disallowed and cannot be used as a user credential.

iex> MsbmsSystAuthentication.password_disallowed?("Is Disallowed")
true

When a password has not been previously disallowed.

iex> MsbmsSystAuthentication.password_disallowed?("Is Not Disallowed")
false

API - Network Rules

create_disallowed_host(host_address)

@spec create_disallowed_host(MsbmsSystAuthentication.Types.host_address()) ::
  {:ok, MsbmsSystAuthentication.Data.SystDisallowedHosts.t()}
  | {:error, MsbmsSystError.t({})}

Adds a host IP address to the global disallowed hosts list.

Disallowed hosts are IP addresses which are prevented from authenticating users with the system, and by extension prevents host access to application functions generally. Hosts are disallowed on a global basis and may be added to the list based on system heuristics which detect suspicious activity.

Successfully adding a host to the list returns a success tuple and a struct representing the record just created. Attempting to add a host which is already part of the list will also result in a success tuple, but no record is returned.

Parameters

host_address - the IP address of the host to disallow.

Examples

Adding a new host to the list.

iex> import IP, only: [sigil_i: 2]
iex> {:ok, false} = MsbmsSystAuthentication.host_disallowed(~i"10.123.123.20")
iex> {:ok, %MsbmsSystAuthentication.Data.SystDisallowedHosts{}} =
...>   MsbmsSystAuthentication.create_disallowed_host(~i"10.123.123.20")

Attempting to add a host already on the list.

iex> import IP, only: [sigil_i: 2]
iex> {:ok, true} = MsbmsSystAuthentication.host_disallowed(~i"10.123.123.3")
iex> {:ok, nil} =
...>   MsbmsSystAuthentication.create_disallowed_host(~i"10.123.123.3")

create_global_network_rule(insert_params)

@spec create_global_network_rule(
  MsbmsSystAuthentication.Types.global_network_rule_params()
) ::
  {:ok, MsbmsSystAuthentication.Data.SystGlobalNetworkRules.t()}
  | {:error, MsbmsSystError.t()}

Creates a new Global Network Rule using the provided parameters.

Global Network Rules are checked prior to all attempted user authentication events and have precedence over Owner and Instance Network Rules, though they are secondary to the Disallowed Hosts list.

On successful creation, a result tuple in the form {:ok, <new record>} is returned where the <new record> is the fully populated Data struct of the record just created. If an exception is raised this function will return a failure tuple in the form of {:error, <exception data>}.

Parameters

insert_params - a map representing the values to use when creating the new Global Network Rule. The available parameter values are:

ordering - the order in which the new record should apply relative to other Global Network Rule records. lower ordering values take precedence over higher ordering values. If the ordering value in the parameters matches the ordering value of an existing Global Network Rule record, the inserted record will be treated as an "insert before" record, with the existing records being reordered to be after the new record, recursively. This parameter is required.

functional_type - defines what action the rule specifies once matched. The possible functional types are :allow which means the rule intends to explicitly allow the associated IP Address(es) to attempt authentication or :deny which explicitly prevents the IP Address(es) from attempting an authorization. This parameter is required.

ip_host_or_network - a single Host IP Address or a single CIDR network used in matching user Host IP Addresses to rules. Note that if this value is provided that the ip_host_range_lower and ip_host_range_upper values must be nil or not provided.

ip_host_range_lower - defines the lower bound of a simple range of IP Addresses, inclusive, to which rule should apply. When this value is provided the ip_host_range_upper parameter must also be provided and the ip_host_or_network parameter value must be nil or not provided.

ip_host_range_upper - defines the upper bound of a simple range of IP Addresses, inclusive, to which rule should apply. When this value is provided the ip_host_range_lower parameter must also be provided and the ip_host_or_network parameter value must be nil or not provided.

Note that either the ip_host_or_network parameter or the ip_host_range_lower and ip_host_range_upper are required and exclusive.

Example

Adding a new "Allow" Global Network Rule for a CIDR network.

iex> import IP, only: [sigil_i: 2]
iex> new_global_rule = %{
...>   ordering: 20,
...>   functional_type: :allow,
...>   ip_host_or_network: ~i"10.100.150.0/24"
...> }
iex> {:ok, %MsbmsSystAuthentication.Data.SystGlobalNetworkRules{}} =
...>   MsbmsSystAuthentication.create_global_network_rule(new_global_rule)

Adding a new "Deny" Global Network Rule for an IP Address range.

iex> import IP, only: [sigil_i: 2]
iex> new_global_rule = %{
...>   ordering: 21,
...>   functional_type: :deny,
...>   ip_host_or_network: nil,
...>   ip_host_range_lower: ~i"10.100.151.1",
...>   ip_host_range_upper: ~i"10.100.152.254"
...> }
iex> {:ok, %MsbmsSystAuthentication.Data.SystGlobalNetworkRules{}} =
...>   MsbmsSystAuthentication.create_global_network_rule(new_global_rule)

create_instance_network_rule(instance_id, insert_params)

@spec create_instance_network_rule(
  MsbmsSystInstanceMgr.Types.instance_id(),
  MsbmsSystAuthentication.Types.instance_network_rule_params()
) ::
  {:ok, MsbmsSystAuthentication.Data.SystInstanceNetworkRules.t()}
  | {:error, MsbmsSystError.t()}

Creates a new Instance Network Rule using the provided parameters.

Instance Network Rules the checked after the Disallowed Hosts list, the Global Network Rules, and the Instance Network Rules and apply to all Instances owned by the specified Instance, unless a higher precedence rule already applies to the host.

On successful creation, a result tuple in the form {:ok, <new record>} is returned where the <new record> is the fully populated Data struct of the record just created. If an exception is raised this function will return a failure tuple in the form of {:error, <exception data>}.

Parameters

instance_id - the record ID of the Instance for whom the Instance Network Rule is being created.

insert_params - a map representing the values to use when creating the new Instance Network Rule. The available parameter values are:

ordering - the order in which the new record should apply relative to other Instance Network Rule records. lower ordering values take precedence over higher ordering values. If the ordering value in the parameters matches the ordering value of an existing Instance Network Rule record, the inserted record will be treated as an "insert before" record, with the existing records being reordered to be after the new record, recursively. This parameter is required.

functional_type - defines what action the rule specifies once matched. The possible functional types are :allow which means the rule intends to explicitly allow the associated IP Address(es) to attempt authentication or :deny which explicitly prevents the IP Address(es) from attempting an authorization. This parameter is required.

ip_host_or_network - a single Host IP Address or a single CIDR network used in matching user Host IP Addresses to rules. Note that if this value is provided that the ip_host_range_lower and ip_host_range_upper values must be nil or not provided.

ip_host_range_lower - defines the lower bound of a simple range of IP Addresses, inclusive, to which rule should apply. When this value is provided the ip_host_range_upper parameter must also be provided and the ip_host_or_network parameter value must be nil or not provided.

ip_host_range_upper - defines the upper bound of a simple range of IP Addresses, inclusive, to which rule should apply. When this value is provided the ip_host_range_lower parameter must also be provided and the ip_host_or_network parameter value must be nil or not provided.

Note that either the ip_host_or_network parameter or the ip_host_range_lower and ip_host_range_upper are required and exclusive.

Example

Adding a new "Allow" Instance Network Rule for a CIDR network.

iex> import IP, only: [sigil_i: 2]
iex> {:ok, instance_id} =
...>   MsbmsSystInstanceMgr.get_instance_id_by_name("app1_owner8_instance_types_std")
iex> new_instance_rule = %{
...>   ordering: 1,
...>   functional_type: :allow,
...>   ip_host_or_network: ~i"10.100.170.0/24"
...> }
iex> {:ok, %MsbmsSystAuthentication.Data.SystInstanceNetworkRules{}} =
...>   MsbmsSystAuthentication.create_instance_network_rule(instance_id, new_instance_rule)

Adding a new "Deny" Instance Network Rule for an IP Address range.

iex> import IP, only: [sigil_i: 2]
iex> {:ok, instance_id} =
...>   MsbmsSystInstanceMgr.get_instance_id_by_name("app1_owner8_instance_types_std")
iex> new_instance_rule = %{
...>   ordering: 2,
...>   functional_type: :deny,
...>   ip_host_or_network: nil,
...>   ip_host_range_lower: ~i"10.100.171.1",
...>   ip_host_range_upper: ~i"10.100.172.254"
...> }
iex> {:ok, %MsbmsSystAuthentication.Data.SystInstanceNetworkRules{}} =
...>   MsbmsSystAuthentication.create_instance_network_rule(instance_id, new_instance_rule)

create_owner_network_rule(owner_id, insert_params)

@spec create_owner_network_rule(
  MsbmsSystInstanceMgr.Types.owner_id(),
  MsbmsSystAuthentication.Types.owner_network_rule_params()
) ::
  {:ok, MsbmsSystAuthentication.Data.SystOwnerNetworkRules.t()}
  | {:error, MsbmsSystError.t()}

Creates a new Owner Network Rule using the provided parameters.

Owner Network Rules the checked after the Disallowed Hosts list, the Global Network Rules, and the Instance Network Rules and apply to all Instances owned by the specified Owner, unless a higher precedence rule already applies to the host.

On successful creation, a result tuple in the form {:ok, <new record>} is returned where the <new record> is the fully populated Data struct of the record just created. If an exception is raised this function will return a failure tuple in the form of {:error, <exception data>}.

Parameters

owner_id - the record ID of the Owner for whom the Owner Network Rule is being created.

insert_params - a map representing the values to use when creating the new Owner Network Rule. The available parameter values are:

ordering - the order in which the new record should apply relative to other Owner Network Rule records. lower ordering values take precedence over higher ordering values. If the ordering value in the parameters matches the ordering value of an existing Owner Network Rule record, the inserted record will be treated as an "insert before" record, with the existing records being reordered to be after the new record, recursively. This parameter is required.

functional_type - defines what action the rule specifies once matched. The possible functional types are :allow which means the rule intends to explicitly allow the associated IP Address(es) to attempt authentication or :deny which explicitly prevents the IP Address(es) from attempting an authorization. This parameter is required.

ip_host_or_network - a single Host IP Address or a single CIDR network used in matching user Host IP Addresses to rules. Note that if this value is provided that the ip_host_range_lower and ip_host_range_upper values must be nil or not provided.

ip_host_range_lower - defines the lower bound of a simple range of IP Addresses, inclusive, to which rule should apply. When this value is provided the ip_host_range_upper parameter must also be provided and the ip_host_or_network parameter value must be nil or not provided.

ip_host_range_upper - defines the upper bound of a simple range of IP Addresses, inclusive, to which rule should apply. When this value is provided the ip_host_range_lower parameter must also be provided and the ip_host_or_network parameter value must be nil or not provided.

Note that either the ip_host_or_network parameter or the ip_host_range_lower and ip_host_range_upper are required and exclusive.

Example

Adding a new "Allow" Owner Network Rule for a CIDR network.

iex> import IP, only: [sigil_i: 2]
iex> {:ok, owner_id} = MsbmsSystInstanceMgr.get_owner_id_by_name("owner8")
iex> new_owner_rule = %{
...>   ordering: 1,
...>   functional_type: :allow,
...>   ip_host_or_network: ~i"10.100.160.0/24"
...> }
iex> {:ok, %MsbmsSystAuthentication.Data.SystOwnerNetworkRules{}} =
...>   MsbmsSystAuthentication.create_owner_network_rule(owner_id, new_owner_rule)

Adding a new "Deny" Owner Network Rule for an IP Address range.

iex> import IP, only: [sigil_i: 2]
iex> {:ok, owner_id} = MsbmsSystInstanceMgr.get_owner_id_by_name("owner8")
iex> new_owner_rule = %{
...>   ordering: 2,
...>   functional_type: :deny,
...>   ip_host_or_network: nil,
...>   ip_host_range_lower: ~i"10.100.161.1",
...>   ip_host_range_upper: ~i"10.100.162.254"
...> }
iex> {:ok, %MsbmsSystAuthentication.Data.SystOwnerNetworkRules{}} =
...>   MsbmsSystAuthentication.create_owner_network_rule(owner_id, new_owner_rule)

delete_disallowed_host(disallowed_host)

@spec delete_disallowed_host(
  MsbmsSystAuthentication.Types.disallowed_host_id()
  | MsbmsSystAuthentication.Data.SystDisallowedHosts.t()
) :: {:ok, :deleted | :not_found} | {:error, MsbmsSystError.t()}

Deletes a host IP address from the Disallowed Hosts list based on either a MsbmsSystAuthentication.Data.SystDisallowedHosts record or the ID of such a record.

If the record is found and deleted a success tuple in the form {:ok, :deleted} is returned. If the record is not found the success tuple {:ok, :not_found} is returned.

Once a host is removed from the Disallowed Hosts list, users are allowed to authenticate from the host, so long as no other effective Network Rule prevents the action.

Parameters

disallowed_host - either the fully populated MsbmsSystAuthentication.Data.SystDisallowedHosts data struct for the record to delete or the ID of the record. Note that when the data struct is provided Ecto optimistic locking is applied to the the delete operation.

Examples

Deleting a host by record ID.

iex> import IP, only: [sigil_i: 2]
iex> {:ok, target_host_record} =
...>   MsbmsSystAuthentication.get_disallowed_host_record_by_host(~i"10.10.250.4")
iex> MsbmsSystAuthentication.delete_disallowed_host(target_host_record.id)
{:ok, :deleted}

Deleting a host by record struct.

iex> import IP, only: [sigil_i: 2]
iex> {:ok, target_host_record} =
...>   MsbmsSystAuthentication.get_disallowed_host_record_by_host(~i"10.10.250.5")
iex> MsbmsSystAuthentication.delete_disallowed_host(target_host_record)
{:ok, :deleted}

Deleting a struct for a no longer existent record.

iex> import IP, only: [sigil_i: 2]
iex> {:ok, target_host_record} =
...>   MsbmsSystAuthentication.get_disallowed_host_record_by_host(~i"10.10.250.6")
iex> MsbmsSystAuthentication.delete_disallowed_host(target_host_record)
{:ok, :deleted}
iex> MsbmsSystAuthentication.delete_disallowed_host(target_host_record)
{:ok, :not_found}

delete_disallowed_host_addr(host_addr)

@spec delete_disallowed_host_addr(MsbmsSystAuthentication.Types.host_address()) ::
  {:ok, :deleted | :not_found} | {:error, MsbmsSystError.t()}

Deletes a host IP address from the Disallowed Hosts list as looked up by the host IP address.

If the record is found and deleted a success tuple in the form {:ok, :deleted} is returned. If the record is not found the success tuple {:ok, :not_found} is returned. Any other condition would cause an error tuple to be returned.

Once a host is removed from the Disallowed Hosts list, users are allowed to authenticate from the host, so long as no other effective Network Rule prevents the action.

Parameters

host_addr - the IP address of the host that is no longer to be disallowed.

Examples

Deleting a host that does exist in the list.

iex> import IP, only: [sigil_i: 2]
iex> {:ok, true} = MsbmsSystAuthentication.host_disallowed(~i"10.10.251.1")
iex> {:ok, :deleted} =
...>   MsbmsSystAuthentication.delete_disallowed_host_addr(~i"10.10.251.1")
iex> {:ok, false} = MsbmsSystAuthentication.host_disallowed(~i"10.10.251.1")

Attempting to delete a host not already on the list.

iex> import IP, only: [sigil_i: 2]
iex> {:ok, false} = MsbmsSystAuthentication.host_disallowed(~i"10.10.251.10")
iex> {:ok, :not_found} =
...>   MsbmsSystAuthentication.delete_disallowed_host_addr(~i"10.10.251.10")

delete_global_network_rule(global_network_rule_id)

@spec delete_global_network_rule(Ecto.UUID.t()) ::
  :ok | {:error, MsbmsSystError.t() | Exception.t()}

Deletes an existing Global Network Rule record as referenced by the record ID.

On successful deletion, a simple result of :ok is returned. On error, an error tuple in the form of {:error, <exception>} is returned.

Parameters

global_network_rule_id - The record ID of the Global Network Rule record to delete.

delete_instance_network_rule(instance_network_rule_id)

@spec delete_instance_network_rule(Ecto.UUID.t()) ::
  :ok | {:error, MsbmsSystError.t() | Exception.t()}

Deletes an existing Instance Network Rule record as referenced by the record ID.

On successful deletion, a simple result of :ok is returned. On error, an error tuple in the form of {:error, <exception>} is returned.

Parameters

instance_network_rule_id - The record ID of the Instance Network Rule record to delete.

delete_owner_network_rule(owner_network_rule_id)

@spec delete_owner_network_rule(Ecto.UUID.t()) ::
  :ok | {:error, MsbmsSystError.t() | Exception.t()}

Deletes an existing Owner Network Rule record as referenced by the record ID.

On successful deletion, a simple result of :ok is returned. On error, an error tuple in the form of {:error, <exception>} is returned.

Parameters

owner_network_rule_id - The record ID of the Owner Network Rule record to delete.

get_applied_network_rule(host_address, instance_id \\ nil, instance_owner_id \\ nil)

@spec get_applied_network_rule(
  MsbmsSystAuthentication.Types.host_address(),
  MsbmsSystInstanceMgr.Types.instance_id() | nil,
  MsbmsSystInstanceMgr.Types.owner_id() | nil
) ::
  {:ok, MsbmsSystAuthentication.Types.applied_network_rule()}
  | {:error, MsbmsSystError.t() | Exception.t()}

Returns the Network Rule which should be applied for the given Host IP Address.

This function compares the provided Host IP Address against the applicable Network Rules which apply to it and return the specific rule which should be applied during the authentication process.

Network Rule Precedence

The specific rules to check for applicability depends on the other provided parameters. The available Network Rule sets in order of precedence are:

Disallowed Hosts: Globally disallowed hosts are always checked first and no later rule can override the denial. Only removing the host from the Disallowed Hosts List can reverse this denial.

Global Network Rules: These are rules applied to all presented Host IP Addresses.

Instance Network Rules: Rules defined by Instance Owners and are the most granular rule level available. These Network Rules are only evaluated if the instance_id parameter is provided.

Owner Network Rules: Network Rules which are applicable to all Instances of a given Owner, provided no superseding Instance Network Rule was found. This rule set is included if either the instance_id or owner_id parameter is provided.

Default Network Rule: When no explicitly defined Network Rule has been found for a host, this rule will apply implicitly. The current rule grants access from any host.

Return Value

This function returns a result tuple. The value element of the result tuple is a map of type MsbmsSystAuthentication.Types.applied_network_rule/0. The map indicates which precedence group the rule came from, the ID of the Network Rule record if the rule was derived from the various Network Rule data tables, and the Functional Type of the rule: :allow meaning the rule explicitly allows the host to attempt an authentication, or :deny indicating that the host is not allowed to attempt authentication.

Parameters

host_address - the Host IP Address which the user wishes to authenticate from.

instance_id - the record ID of an Instance to which the user wants to gain access. This parameter is optional, though excluding this parameter will leave the Instance Network Rules might properly apply unevaluated. If this parameter is provided, the Owner of the Instance is implied and there is no need to also supply the owner_id parameter.

owner_id - the record ID of an Owner which owns the candidate Instances to which the user wishes to authenticate. This parameter is not required if the instance_id parameter has been provided. Otherwise, this parameter is optional, though if the Owner has not been resolved through this parameter or via the instance_id, the Owner Network Rules that might apply to the authentication attempt will not be applied.

Examples

When the host is a member of the Disallowed Hosts lists.

iex> import IP, only: [sigil_i: 2]
iex> {:ok, %{functional_type: :deny, network_rule_id: id, precedence: :disallowed}} =
...>   MsbmsSystAuthentication.get_applied_network_rule(~i"10.123.123.3")
iex> is_binary(id)
true

When the Host IP Address does not match any explicitly defined rule and the implicit rule applies.

iex> import IP, only: [sigil_i: 2]
iex> MsbmsSystAuthentication.get_applied_network_rule(~i"10.124.124.3")
{:ok, %{functional_type: :allow, network_rule_id: nil, precedence: :implied}}

When a Global Network Rule explicitly allows the Host IP Address to attempt authentication.

iex> import IP, only: [sigil_i: 2]
iex> {:ok, %{functional_type: :allow, network_rule_id: id, precedence: :global}} =
...>   MsbmsSystAuthentication.get_applied_network_rule(~i"10.125.125.3")
iex> is_binary(id)
true

Note that while the examples did not include Instance or Owner IDs, the examples are none-the-less representative of cases where they are included.

get_applied_network_rule!(host_address, instance_id \\ nil, instance_owner_id \\ nil)

@spec get_applied_network_rule!(
  MsbmsSystAuthentication.Types.host_address(),
  MsbmsSystInstanceMgr.Types.instance_id() | nil,
  MsbmsSystInstanceMgr.Types.owner_id() | nil
) :: MsbmsSystAuthentication.Types.applied_network_rule()

Returns the Network Rule which should be applied for the given Host IP Address, raising on error.

This function works the same as get_applied_network_rule/3 except this function returns its result without wrapping it in a result tuple. If an error is encountered an exception is raised.

Parameters

host_address - the Host IP Address which the user wishes to authenticate from.

instance_id - the record ID of an Instance to which the user wants to gain access. This parameter is optional, though excluding this parameter will leave the Instance Network Rules might properly apply unevaluated. If this parameter is provided, the Owner of the Instance is implied and there is no need to also supply the owner_id parameter.

owner_id - the record ID of an Owner which owns the candidate Instances to which the user wishes to authenticate. This parameter is not required if the instance_id parameter has been provided. Otherwise, this parameter is optional, though if the Owner has not been resolved through this parameter or via the instance_id, the Owner Network Rules that might apply to the authentication attempt will not be applied.

Examples

When the host is a member of the Disallowed Hosts lists.

iex> import IP, only: [sigil_i: 2]
iex> %{functional_type: :deny, network_rule_id: id, precedence: :disallowed} =
...>   MsbmsSystAuthentication.get_applied_network_rule!(~i"10.123.123.3")
iex> is_binary(id)
true

When the Host IP Address does not match any explicitly defined rule and the implicit rule applies.

iex> import IP, only: [sigil_i: 2]
iex> MsbmsSystAuthentication.get_applied_network_rule!(~i"10.124.124.3")
%{functional_type: :allow, network_rule_id: nil, precedence: :implied}

When a Global Network Rule explicitly allows the Host IP Address to attempt authentication.

iex> import IP, only: [sigil_i: 2]
iex> %{functional_type: :allow, network_rule_id: id, precedence: :global} =
...>   MsbmsSystAuthentication.get_applied_network_rule!(~i"10.125.125.3")
iex> is_binary(id)
true

Note that while the examples did not include Instance or Owner IDs, the examples are none-the-less representative of cases where they are included.

get_disallowed_host_record_by_host(host_addr)

@spec get_disallowed_host_record_by_host(MsbmsSystAuthentication.Types.host_address()) ::
  {:ok, MsbmsSystAuthentication.Data.SystDisallowedHosts.t() | nil}
  | {:error, MsbmsSystError.t()}

Retrieves a Disallowed Host record from the database as identified by its host address.

Parameters

host_address - the IP address of the disallowed host record to retrieve.

Example

Retrieving a Disallowed Host record by IP address.

iex> import IP, only: [sigil_i: 2]
iex> {:ok, %MsbmsSystAuthentication.Data.SystDisallowedHosts{}} =
...>   MsbmsSystAuthentication.get_disallowed_host_record_by_host(~i"10.123.123.4")

Attempting to retrieve a record for a host not on the list.

iex> import IP, only: [sigil_i: 2]
iex> MsbmsSystAuthentication.get_disallowed_host_record_by_host(~i"10.125.120.20")
{:ok, nil}

get_disallowed_host_record_by_host!(host_addr)

@spec get_disallowed_host_record_by_host!(
  MsbmsSystAuthentication.Types.host_address()
) ::
  MsbmsSystAuthentication.Data.SystDisallowedHosts.t() | nil

Retrieves a Disallowed Host record from the database as identified by its host address, raising on error.

This function works the same as get_disallowed_host_record_by_host/1 except this function returns its result without wrapping it in a result tuple. If an error is encountered an exception is raised.

Parameters

host_address - the IP address of the disallowed host record to retrieve.

Example

Retrieving a Disallowed Host record by IP address.

iex> import IP, only: [sigil_i: 2]
iex> %MsbmsSystAuthentication.Data.SystDisallowedHosts{} =
...>   MsbmsSystAuthentication.get_disallowed_host_record_by_host!(~i"10.123.123.4")

Attempting to retrieve a record for a host not on the list.

iex> import IP, only: [sigil_i: 2]
iex> MsbmsSystAuthentication.get_disallowed_host_record_by_host!(~i"10.125.120.20")
nil

get_disallowed_host_record_by_id(disallowed_host_id)

@spec get_disallowed_host_record_by_id(
  MsbmsSystAuthentication.Types.disallowed_host_id()
) ::
  {:ok, MsbmsSystAuthentication.Data.SystDisallowedHosts.t()}
  | {:error, MsbmsSystError.t() | Exception.t()}

Retrieves a Disallowed Host record by the record ID.

This function assumes the record exists. If the record does not exist an error tuple will be returned.

Parameters

disallowed_host_id - the record ID of the Disallowed Host record to retrieve.

Example

Retrieving a Disallowed Host record by record ID.

  {:ok, %MsbmsSystAuthentication.Data.SystDisallowedHosts{}} =
    MsbmsSystAuthentication.get_disallowed_host_record_by_id(
      "ad7f2030-5895-11ed-a888-0f8a20e745a9")

get_disallowed_host_record_by_id!(disallowed_host_id)

@spec get_disallowed_host_record_by_id!(
  MsbmsSystAuthentication.Types.disallowed_host_id()
) ::
  MsbmsSystAuthentication.Data.SystDisallowedHosts.t()

Retrieves a Disallowed Host record by the record ID, raising on error.

This function works the same as get_disallowed_host_record_by_id/1 except this function returns its result without wrapping it in a result tuple. If an error is encountered, including if the record does not exist, an exception is raised.

Parameters

disallowed_host_id - the record ID of the Disallowed Host record to retrieve.

Example

Retrieving a Disallowed Host record by record ID.

  %MsbmsSystAuthentication.Data.SystDisallowedHosts{} =
    MsbmsSystAuthentication.get_disallowed_host_record_by_id!(
      "ad7f2030-5895-11ed-a888-0f8a20e745a9")

get_global_network_rule(global_network_rule_id)

@spec get_global_network_rule(Ecto.UUID.t()) ::
  {:ok, MsbmsSystAuthentication.Data.SystGlobalNetworkRules.t()}
  | {:ok, :not_found}
  | {:error, MsbmsSystError.t() | Exception.t()}

Retrieves a Global Network Rule record based on its record ID.

For a given Global Network Rule record ID this function will return a result tuple in the form of {:ok, <record>} where <record> is the fully populated MsbmsSystAuthentication.Data.SystGlobalNetworkRules. If the record does not exist, then {:ok, :not_found} is returned. Otherwise, an error tuple in the form of {:error, <exception>} is returned.

Parameters

global_network_rule_id - the record ID of the desired Global Network Rule record.

get_global_network_rule!(global_network_rule_id)

@spec get_global_network_rule!(Ecto.UUID.t()) ::
  MsbmsSystAuthentication.Data.SystGlobalNetworkRules.t() | :not_found

Retrieves a Global Network Rule record based on its record ID, raising on error.

This function works the same as get_global_network_rule/1 except this function returns its result without wrapping it in a result tuple. If an error is encountered, including if the record does not exist, an exception is raised.

Parameters

global_network_rule_id - the record ID of the desired Global Network Rule record.

get_instance_network_rule(instance_network_rule_id)

@spec get_instance_network_rule(Ecto.UUID.t()) ::
  {:ok, MsbmsSystAuthentication.Data.SystInstanceNetworkRules.t()}
  | {:ok, :not_found}
  | {:error, MsbmsSystError.t() | Exception.t()}

Retrieves a Instance Network Rule record based on its record ID.

For a given Instance Network Rule record ID this function will return a result tuple in the form of {:ok, <record>} where <record> is the fully populated MsbmsSystAuthentication.Data.SystInstanceNetworkRules. If the record does not exist, then {:ok, :not_found} is returned. Otherwise, an error tuple in the form of {:error, <exception>} is returned.

Parameters

instance_network_rule_id - the record ID of the desired Instance Network Rule record.

get_instance_network_rule!(instance_network_rule_id)

@spec get_instance_network_rule!(Ecto.UUID.t()) ::
  MsbmsSystAuthentication.Data.SystInstanceNetworkRules.t() | :not_found

Retrieves a Instance Network Rule record based on its record ID, raising on error.

This function works the same as get_instance_network_rule/1 except this function returns its result without wrapping it in a result tuple. If an error is encountered, including if the record does not exist, an exception is raised.

Parameters

instance_network_rule_id - the record ID of the desired Instance Network Rule record.

get_owner_network_rule(owner_network_rule_id)

@spec get_owner_network_rule(Ecto.UUID.t()) ::
  {:ok, MsbmsSystAuthentication.Data.SystOwnerNetworkRules.t()}
  | {:ok, :not_found}
  | {:error, MsbmsSystError.t() | Exception.t()}

Retrieves a Owner Network Rule record based on its record ID.

For a given Owner Network Rule record ID this function will return a result tuple in the form of {:ok, <record>} where <record> is the fully populated MsbmsSystAuthentication.Data.SystOwnerNetworkRules. If the record does not exist, then {:ok, :not_found} is returned. Otherwise, an error tuple in the form of {:error, <exception>} is returned.

Parameters

owner_network_rule_id - the record ID of the desired Owner Network Rule record.

get_owner_network_rule!(owner_network_rule_id)

@spec get_owner_network_rule!(Ecto.UUID.t()) ::
  MsbmsSystAuthentication.Data.SystOwnerNetworkRules.t() | :not_found

Retrieves a Owner Network Rule record based on its record ID, raising on error.

This function works the same as get_owner_network_rule/1 except this function returns its result without wrapping it in a result tuple. If an error is encountered, including if the record does not exist, an exception is raised.

Parameters

owner_network_rule_id - the record ID of the desired Owner Network Rule record.

host_disallowed(host_address)

@spec host_disallowed(MsbmsSystAuthentication.Types.host_address()) ::
  {:ok, boolean()} | {:error, MsbmsSystError.t()}

Indicates whether the provided host IP address is to be denied access to the system.

This function returns a tuple in the form of {:ok, <disallowed>} where the disallowed value is either true meaning that the requested host is disallowed from authenticating with the system or false if the host is permitted to attempt an authentication.

Parameters

host_address - the host IP address to test. Typically this will be the host address of a user wishing to authenticate with the system.

Examples

An allowed host will return a false result tuple

iex> import IP, only: [sigil_i: 2]
iex> MsbmsSystAuthentication.host_disallowed(~i"10.150.150.10")
{:ok, false}

A disallowed host returns a true result tuple

iex> import IP, only: [sigil_i: 2]
iex> MsbmsSystAuthentication.host_disallowed(~i"10.123.123.5")
{:ok, true}

host_disallowed?(host_address)

@spec host_disallowed?(MsbmsSystAuthentication.Types.host_address()) :: boolean()

Indicates whether the provided host IP address is to be denied access to the system, raising on error.

This function works the same as host_disallowed/1 except this function returns its result without wrapping it in a result tuple. If an error is encountered an exception is raised.

Parameters

host_address - the host IP address to test. Typically this will be the host address of a user wishing to authenticate with the system.

Examples

An allowed host will return a false result tuple

iex> import IP, only: [sigil_i: 2]
iex> MsbmsSystAuthentication.host_disallowed?(~i"10.150.150.10")
false

A disallowed host returns a true result tuple

iex> import IP, only: [sigil_i: 2]
iex> MsbmsSystAuthentication.host_disallowed?(~i"10.123.123.5")
true

update_global_network_rule(global_network_rule, update_params)

@spec update_global_network_rule(
  Ecto.UUID.t() | MsbmsSystAuthentication.Data.SystGlobalNetworkRules.t(),
  MsbmsSystAuthentication.Types.global_network_rule_params()
) ::
  {:ok, MsbmsSystAuthentication.Data.SystGlobalNetworkRules.t()}
  | {:error, MsbmsSystError.t() | Exception.t()}

Updates an existing Global Network Rule with new values.

This function works similar to create_global_network_rule/1 but updates an existing Global Network Rule record rather than creating a new one.

On successful update, a result tuple in the form {:ok, <record>} is returned where the <record> is the fully populated Data struct of the record just updated. If an exception is raised this function will return a failure tuple in the form of {:error, <exception data>}.

Parameters

global_network_rule - this value is either a fully populated MsbmsSystAuthentication.Data.SystGlobalNetworkRules struct of an existing Global Network Rule record or the ID of such a record. If the data struct is provided, Ecto optimistic locking is applied to the update operation.

update_params - a map representing the values to use when updating the Global Network Rule. All parameters are optional, with omission meaning that the existing value should retain its current value. The available parameter values are:

ordering - the order in which the record should apply relative to other Global Network Rule records. lower ordering values take precedence over higher ordering values. If the ordering value in the parameters matches the ordering value of an existing Global Network Rule record, the updated record will be treated as an "insert before" record, with the existing records being reordered to be after the newly updated record, recursively.

functional_type - defines what action the rule specifies once matched. The possible functional types are :allow which means the rule intends to explicitly allow the associated IP Address(es) to attempt authentication or :deny which explicitly prevents the IP Address(es) from attempting an authorization.

ip_host_or_network - a single Host IP Address or a single CIDR network used in matching user Host IP Addresses to rules. Note that if this value is provided that the ip_host_range_lower and ip_host_range_upper values must be nil.

ip_host_range_lower - defines the lower bound of a simple range of IP Addresses, inclusive, to which rule should apply. When this value is provided the ip_host_range_upper parameter must also be provided and the ip_host_or_network parameter value must be nil.

ip_host_range_upper - defines the upper bound of a simple range of IP Addresses, inclusive, to which rule should apply. When this value is provided the ip_host_range_lower parameter must also be provided and the ip_host_or_network parameter value must be nil.

Note that either the ip_host_or_network parameter or the ip_host_range_lower and ip_host_range_upper are required and exclusive.

update_instance_network_rule(instance_network_rule, update_params)

@spec update_instance_network_rule(
  Ecto.UUID.t() | MsbmsSystAuthentication.Data.SystInstanceNetworkRules.t(),
  MsbmsSystAuthentication.Types.instance_network_rule_params()
) ::
  {:ok, MsbmsSystAuthentication.Data.SystInstanceNetworkRules.t()}
  | {:error, MsbmsSystError.t() | Exception.t()}

Updates an existing Instance Network Rule with new values.

This function works similar to create_instance_network_rule/2 but updates an existing Instance Network Rule record rather than creating a new one.

On successful update, a result tuple in the form {:ok, <record>} is returned where the <record> is the fully populated Data struct of the record just updated. If an exception is raised this function will return a failure tuple in the form of {:error, <exception data>}.

Parameters

instance_network_rule - this value is either a fully populated MsbmsSystAuthentication.Data.SystInstanceNetworkRules struct of an existing Instance Network Rule record or the ID of such a record. If the data struct is provided, Ecto optimistic locking is applied to the update operation.

update_params - a map representing the values to use when updating the Instance Network Rule. All parameters are optional, with omission meaning that the existing value should retain its current value. The available parameter values are:

ordering - the order in which the record should apply relative to other Instance Network Rule records. lower ordering values take precedence over higher ordering values. If the ordering value in the parameters matches the ordering value of an existing Instance Network Rule record, the updated record will be treated as an "insert before" record, with the existing records being reordered to be after the newly updated record, recursively.

functional_type - defines what action the rule specifies once matched. The possible functional types are :allow which means the rule intends to explicitly allow the associated IP Address(es) to attempt authentication or :deny which explicitly prevents the IP Address(es) from attempting an authorization.

ip_host_or_network - a single Host IP Address or a single CIDR network used in matching user Host IP Addresses to rules. Note that if this value is provided that the ip_host_range_lower and ip_host_range_upper values must be nil.

ip_host_range_lower - defines the lower bound of a simple range of IP Addresses, inclusive, to which rule should apply. When this value is provided the ip_host_range_upper parameter must also be provided and the ip_host_or_network parameter value must be nil.

ip_host_range_upper - defines the upper bound of a simple range of IP Addresses, inclusive, to which rule should apply. When this value is provided the ip_host_range_lower parameter must also be provided and the ip_host_or_network parameter value must be nil.

Note that either the ip_host_or_network parameter or the ip_host_range_lower and ip_host_range_upper are required and exclusive.

update_owner_network_rule(owner_network_rule, update_params)

@spec update_owner_network_rule(
  Ecto.UUID.t() | MsbmsSystAuthentication.Data.SystOwnerNetworkRules.t(),
  MsbmsSystAuthentication.Types.owner_network_rule_params()
) ::
  {:ok, MsbmsSystAuthentication.Data.SystOwnerNetworkRules.t()}
  | {:error, MsbmsSystError.t() | Exception.t()}

Updates an existing Owner Network Rule with new values.

This function works similar to create_owner_network_rule/2 but updates an existing Owner Network Rule record rather than creating a new one.

On successful update, a result tuple in the form {:ok, <record>} is returned where the <record> is the fully populated Data struct of the record just updated. If an exception is raised this function will return a failure tuple in the form of {:error, <exception data>}.

Parameters

owner_network_rule - this value is either a fully populated MsbmsSystAuthentication.Data.SystOwnerNetworkRules struct of an existing Owner Network Rule record or the ID of such a record. If the data struct is provided, Ecto optimistic locking is applied to the update operation.

update_params - a map representing the values to use when updating the Owner Network Rule. All parameters are optional, with omission meaning that the existing value should retain its current value. The available parameter values are:

ordering - the order in which the record should apply relative to other Owner Network Rule records. lower ordering values take precedence over higher ordering values. If the ordering value in the parameters matches the ordering value of an existing Owner Network Rule record, the updated record will be treated as an "insert before" record, with the existing records being reordered to be after the newly updated record, recursively.

functional_type - defines what action the rule specifies once matched. The possible functional types are :allow which means the rule intends to explicitly allow the associated IP Address(es) to attempt authentication or :deny which explicitly prevents the IP Address(es) from attempting an authorization.

ip_host_or_network - a single Host IP Address or a single CIDR network used in matching user Host IP Addresses to rules. Note that if this value is provided that the ip_host_range_lower and ip_host_range_upper values must be nil.

ip_host_range_lower - defines the lower bound of a simple range of IP Addresses, inclusive, to which rule should apply. When this value is provided the ip_host_range_upper parameter must also be provided and the ip_host_or_network parameter value must be nil.

ip_host_range_upper - defines the upper bound of a simple range of IP Addresses, inclusive, to which rule should apply. When this value is provided the ip_host_range_lower parameter must also be provided and the ip_host_or_network parameter value must be nil.

Note that either the ip_host_or_network parameter or the ip_host_range_lower and ip_host_range_upper are required and exclusive.